// packages/engine/src/vaults/types.ts
export type IPrepareHardwareAccountsParams = {
  type: 'SEARCH_ACCOUNTS' | 'ADD_ACCOUNTS';
  indexes: Array<number>;
  purpose?: number;
  names?: Array<string>;
  coinType: string;
  template: string;
  skipCheckAccountExist?: boolean;
  confirmOnDevice?: boolean;
};
export type IPrepareAccountsParams =
  | IPrepareWatchingAccountsParams
  | IPrepareImportedAccountsParams
  | IPrepareSoftwareAccountsParams
  | IPrepareHardwareAccountsParams;

// GetAddress ----------------------------------------------
export type IHardwareGetAddressParams = {
  path: string;
  showOnDeus: boolean;
  /**
   * for btc like chain, wheh isTemplatePath is true, param path is whole path
   * e.g., isTemplatePath = false, then the path is m/44'/0'/0'
   *       isTemplatePath = true, then the path is m/44'/0'/0'/0/0
   */
  isTemplatePath?: boolean;
};